# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Thrift Store",
        "version": "1.0.0"
    },
    "paths": {
        "/api/expenses": {
            "get": {
                "summary": "Get list of expenses",
                "operationId": "9b702aa9314d471a9de3a121e5183c88",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "user_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "expense_date": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "Y-m-d"
                                            },
                                            "amount": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 150
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Coffee Break"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "summary": "Create a new expense",
                "operationId": "a764a04e6d047ab02a68f06c706e67f3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "description",
                                    "amount",
                                    "user_id",
                                    "expense_date"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "expense_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "Y-m-d"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 150
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Coffee Break"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Expense created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "expense_date": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "Y-m-d"
                                        },
                                        "amount": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 150
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Coffee Break"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/expenses/{id}": {
            "get": {
                "summary": "Get a specific expense",
                "operationId": "97128dceb5c799772eef314ba1b9b045",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "expense_date": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "Y-m-d"
                                        },
                                        "amount": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 150
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Coffee Break"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Expense not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "summary": "Delete an existing expense",
                "operationId": "68ebff62f3bfb2213565253975db2c52",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Expense deleted successfully"
                    },
                    "404": {
                        "description": "Expense not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "summary": "Update an existing expense",
                "operationId": "0e0b366b918838595f72a88fe371291d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "expense_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "Y-m-d"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 150
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Coffee Break"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Expense updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "expense_date": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "Y-m-d"
                                        },
                                        "amount": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 150
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Coffee Break"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Expense not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Bearer token for Sanctum authentication",
                "name": "bearerAuth",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}